FORGEPLUGINS-217: Fixes compilation errors in Forge 3.7.2.Final+
Motivation
----------
This is caused by FORGE-2759, but as specified in https://github.com/forge/core/#what-scope-should-my-addon-dependencies-be, exposed addons should not be in provided scope.

Modifications
-------------
Since UI classes are exposed to the outside world, it should not be declared as provided.

Result
------
Compilation succeeds and works on all Forge versions.